fix: move plan result into own file
