+++ wandelbots/core/planner.py
@@ -1,7 +1,8 @@
 from typing import Union
 
-from wandelbots import MotionGroup
 from wandelbots.apis import motion as motion_api
+from wandelbots.core.motiongroup import MotionGroup
+from wandelbots.core.plan_result import PlanResult
 from wandelbots.exceptions import PlanningFailedException, PlanningPartialSuccessWarning
 from wandelbots.types import (
     Command,
@@ -19,13 +20,6 @@ from wandelbots.types import (
 from wandelbots.util.logger import _get_logger
 
 CommandType = Union[Command, IOValue]
-
-
-class PlanResult:
-    def __init__(self, plan_response: PlanSuccessfulResponse, io_actions: tuple[SetIO, ...] = ()):
-        self.plan_response = plan_response
-        self.io_actions = io_actions
-        self.motion = plan_response.motion
 
 
 class Planner:
@@ -100,7 +94,7 @@ class Planner:
         move_commands, io_actions = self._resolve_commands(trajectory)
         rae_plan_request = self._create_plan_request(tcp, move_commands, start_joints)
         plan_response = self._plan_with_rae(rae_plan_request)
-        return self.PlanResult(plan_response, io_actions)
+        return PlanResult(plan_response, io_actions)
 
     async def plan_async(
         self, trajectory: list[CommandType], start_joints: list[float], tcp: str = None
@@ -109,7 +103,7 @@ class Planner:
         move_commands, io_actions = self._resolve_commands(trajectory)
         rae_plan_request = self._create_plan_request(tcp, move_commands, start_joints)
         plan_response = await self._plan_with_rae_async(rae_plan_request)
-        return self.PlanResult(plan_response, io_actions)
+        return PlanResult(plan_response, io_actions)
 
     @staticmethod
     def _resolve_commands(trajectory: list[CommandType]) -> tuple[list[Command], tuple[SetIO, ...]]:
